requirements.txt
djangorestframework==3.9.0
djangorestframework-datatables==0.4.0
django-adminlte2==0.4.1
Django==2.1
python-dotenv==0.7.1
social-auth-app-django==3.1.0